Separates inbox tests into multiple files
